refactor(KEN-1054): cut the agreement rule and value validators
The branch grew a per-harness agreement rule across matching lock entries
and validators for the recorded commit, the rendered hash and the repo
spelling. Neither is reachable from what the issue orders: kendex's own
apply writes both values, so only a hand-edited lock reaches the shapes
the validators refused, and the Done-when asks the marker to carry what
the lock recorded, not to adjudicate between installs.

The marker now reads the first matching entry and truncates each value to
seven characters, with the unlocked sentinel where the lock recorded none,
as the branch's first commit had it. The report module's tests are back to
the eight main carries, the split-out tests file is gone with them, the
compat loop is back to the two states the Done-when names, and the Triage
bullet is back to its one line.
