+++ .agents/skills/kendex-issues/SKILL.md
@@ -43,17 +43,8 @@ are a defect.
 
 ## Triage
 
-- **Already fixed** → `source=<repo>@<commit7>` in the report marker is the
-  commit the reporter installed from. In a fetched kendex checkout, close with
-  "fixed in <sha>" when `git merge-base --is-ancestor <commit7> <fix-sha>^`
-  exits 0 — against the fix's parent, so an install sitting ON the fix stays
-  open; that report is the one saying the fix did not work. `source=unlocked`,
-  or a commit `git cat-file -e <commit7>^{commit}` cannot find, dates nothing:
-  triage on the symptom and never close on age. `rendered=<hash7>` fingerprints
-  the bytes that reporter's apply wrote and cannot be recomputed here — the
-  consumer's injected instructions are inside it, so it differs per project and
-  per harness at one commit. It tells two installs apart, never an edited one;
-  a local edit is `kendex verify` in the reporter's tree.
+- **Already fixed** → compare the marker's `source=<repo>@<commit7>` to the fix
+  commit before investigating; older than the fix closes with "fixed in <sha>".
 - **Duplicate** → close, name the canonical issue.
 - **Not a kendex asset** → ownership is the asset's SKILL.md frontmatter
   (`source: kendex`), never its install path. Close with the reason; repost
