+++ changelog.d/changed/KEN-1054.md
@@ -1,1 +1,1 @@
-- `kendex report` stamps the source commit and rendered hash the lock recorded for the reported asset into the issue marker, so triage can date a report; what installs disagree on stamps `unlocked`.
+- `kendex report` stamps the source commit and rendered hash the lock recorded for the reported asset into the issue marker, so triage can date a report; an undated asset stamps `unlocked`.
